[bug fixed] change learning rate args type from int to float

+++ train.py
@@ -23,7 +23,7 @@ if __name__ == '__main__':
     parser.add_argument('-net', type=str, required=True, help='net type')
     parser.add_argument('-w', type=int, default=2, help='number of workers for dataloader')
     parser.add_argument('-b', type=int, default=64, help='batch size for dataloader')
-    parser.add_argument('-lr', type=int, default=0.1, help='initial learning rate')
+    parser.add_argument('-lr', type=float, default=0.1, help='initial learning rate')
     parser.add_argument('-e', type=int, default=120, help='training epoches')
     args = parser.parse_args()
